# align IRI authority component parsing with RFC 3987