Fixes #62

There's definitely performance degradation, about 12% to 50% slower than before #59 landed. But the absolute times still seem low, so hopefully that's acceptable (and it may also be possible to better optimize the new regular expressions).

`benchstat` vs. master@7e92b3eaa72bc6d3466bfdc9fe61f53389fdf47e

    name             old time/op    new time/op    delta
    StrictEmpty-4      14.7ns ± 1%    14.7ns ± 1%       ~     (p=0.238 n=10+10)
    StrictSingle-4     34.3µs ± 4%    49.5µs ± 4%    +43.99%  (p=0.000 n=10+10)
    StrictMany-4        105µs ± 2%     142µs ± 1%    +35.33%  (p=0.000 n=10+9)
    RelaxedEmpty-4     10.0µs ± 1%    22.6µs ± 4%   +126.43%  (p=0.000 n=10+10)
    RelaxedSingle-4    58.1µs ± 2%   143.7µs ± 3%   +147.18%  (p=0.000 n=10+10)
    RelaxedMany-4       156µs ± 1%     371µs ± 6%   +137.41%  (p=0.000 n=9+10)

    name             old alloc/op   new alloc/op   delta
    StrictEmpty-4       0.00B          0.00B            ~     (all equal)
    StrictSingle-4       194B ± 0%      770B ± 4%   +296.70%  (p=0.000 n=9+10)
    StrictMany-4         248B ±16%     1921B ± 3%   +675.93%  (p=0.000 n=10+10)
    RelaxedEmpty-4     3.00B ±100%    15.88B ± 6%   +429.17%  (p=0.000 n=10+8)
    RelaxedSingle-4      333B ± 6%     3865B ± 4%  +1060.89%  (p=0.000 n=10+10)
    RelaxedMany-4        544B ±12%     7601B ± 5%  +1296.14%  (p=0.000 n=10+10)

    name             old allocs/op  new allocs/op  delta
    StrictEmpty-4        0.00           0.00            ~     (all equal)
    StrictSingle-4       2.00 ± 0%      2.00 ± 0%       ~     (all equal)
    StrictMany-4         4.00 ± 0%      4.00 ± 0%       ~     (all equal)
    RelaxedEmpty-4       0.00           0.00            ~     (all equal)
    RelaxedSingle-4      3.00 ± 0%      3.00 ± 0%       ~     (all equal)
    RelaxedMany-4        5.00 ± 0%      5.00 ± 0%       ~     (all equal)