@gibson042 2022-01-24
Upon reconsideration, this is a bit too much in the absence of lookahead/lookbehind assertions.